backend/requirements.txt

Flask
Flask-SQLAlchemy
Flask-Cors
Flask-Migrate
PyJWT
bcrypt
python-dotenv
marshmallow
Flask-Admin
	
